Remove unnecessary shared date functions
