+++ src/modules/billing/lib/charge-period.js
@@ -1,7 +1,26 @@
 'use strict';
 
-const dateHelpers = require('./date-helpers');
+const moment = require('moment');
+const { sortBy, first, last, identity } = require('lodash');
+
 const DateRange = require('../../../lib/models/date-range');
+
+/**
+ * Given an array of dates which can be parsed by Moment,
+ * filters out falsey values and returns a list of moment objects
+ * sorted in ascending date order
+ * @param {Array<String>} arr
+ * @return {Array<Object>}
+ */
+const getSortedDates = arr => sortBy(
+  arr
+    .filter(identity)
+    .map(value => moment(value)),
+  m => m.unix()
+);
+
+const getMinDate = arr => first(getSortedDates(arr));
+const getMaxDate = arr => last(getSortedDates(arr));
 
 /**
  * Gets dates for charge period start and end date functions
@@ -19,10 +38,10 @@ const getDates = (financialYear, chargeVersion, isStartDate = true) => {
   ];
 };
 
-const getChargePeriodStartDate = (financialYear, chargeVersion) => dateHelpers.getMaxDate(
+const getChargePeriodStartDate = (financialYear, chargeVersion) => getMaxDate(
   getDates(financialYear, chargeVersion)).format('YYYY-MM-DD');
 
-const getChargePeriodEndDate = (financialYear, chargeVersion) => dateHelpers.getMinDate(
+const getChargePeriodEndDate = (financialYear, chargeVersion) => getMinDate(
   getDates(financialYear, chargeVersion, false)).format('YYYY-MM-DD');
 
 const getChargePeriod = (financialYear, chargeVersion) => {
